requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-braces==1.13.0
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-cors-headers==3.0.2
django-nose==1.4.6
django-oauth-toolkit==1.2.0
django-timezone-field==3.0
djangorestframework==3.10.0
idna==2.8
kombu==4.6.3
mysqlclient==1.4.2.post1
nose==1.3.7
oauthlib==3.0.2
python-crontab==2.3.8
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
validate-email==1.3
vine==1.3.0
